separated routes from main
